requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy<2.0.0,>=1.21.0
flask>=2.3.0
flask-cors>=4.0.0
pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.25.0
plotly>=5.0.0
pandas>=1.5.0 